requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# MinIO
minio==7.2.0

# Celery
celery==5.3.4
redis==5.0.1

# AI/Computer Vision
insightface==0.7.3
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
onnxruntime==1.16.3
scikit-learn==1.3.2
matplotlib==3.8.2
albumentations==1.3.1

# OCR
paddlepaddle==2.5.2
paddleocr==2.7.3

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-json-logger==2.0.7
aiofiles==23.2.1
